Webscraper that finds radionuclides' symbol, atomic number, mass number, half life, decay constant, gamma energies and intensities.
gamma_search takes in a peak energy and goes to the webpage at
http://nucleardata.nuclear.lu.se/toi/ to input the peak energy. The
program
then searches for the radionuclides that emit gamma energies within 1
keV for gamma energies below 2000 keV and within 4 keV for gamma
energies equal to and above 2000 keV. It then outputs a list of urls for
radionuclides that emit within the gamma energy range and that have an
intensity of at least 1% as well as having a half life of at least 1
hour. These urls are the input for the gamma-extract function.

gamma_extract takes in one url at a time from the gamma_search function
and accesses the webpage of that radionuclide. It then extracts all
gamma energies that have an intensity of at least 1% and also extracts
the corresponding intensities as well. It also extracts the symbol of
the radionuclide (including metastable states) as well as the atomic and
mass numbers. Finally, the half life and decay constant are extracted as
well. gamma_extract returns the gamma energies, corresponding
intensities, symbol, atomic number, mass number, half life (in seconds),
and decay constant (in seconds).