Updated spectrum error checker
Added Bias_Check module that looks at high energy peaks corresponding to
Bi-214 and Tl-208. The module returns a flag if there's a negative value
beyond 2 sigma, where sigma is a net area uncertainty. Flags are
collected for corresponding spectra and a text file is generated. The
text file lists which spectra contain a bias.