Fixed Background Subtraction and Added Isotope Concentration calculations
Removed Background_Subtract and implemented a Peak Finder. The Peak
finder will attempt to find a peak in a strict energy range given by
FWHM. This gives a better estimate of the peak location than the initial
estimate from the isotope energy.

Concentrations are then calculated by comparing the activity to that of
a reference sample. Currently, concentrations are calculated using a
"Conversion Factor".  This needs to be further looked into.