Added calibration function
Added calibration function that finds all the files in the current
directory and checks the current energy calibration using 8 energies
that occur in natural background. If the calibration is unsatisfactory,
a new spectra file is made that contains a corrected calibration curve
that is generated using least squares linear regression.