Implemented sample mass division and adjusted peak measurement calculation
Divided the sample data by the sample mass in kilograms.

Also widened the regions of interest in order to incorporate more of the
peak signal. Regions of interest cover a range where the half-width is
1.3 FWHM. This is roughly 3 sigma and should incorporate 99% of the peak
counts.

Also moved left and right regions 4 FWHM away from the determined peak
region. This should prevent subtracting the peak signal.