Adjusted gross area calculation and added Cs-134 exception
Changed the gross counts calculation such that regions are generated
using channels rather than energy. The given energy input is changed to
a channel that is forced to be an integer. FWHM is also changed to
reflect channels. This ensures that all regions are the same size.

An exception was also made in the compton subtraction for Cs-134. A new
input sub_regions determines whether to use an individual region or both
regions to do the compton subtraction. Calculations were adjusted to
accomadate the use of one or both regions while still keeping the answer
mathematically correct.