Modules for NAA_Analysis.py
isotope_verifier checks if the inputed isotope object is a naturally
occuring isotope and if it is will return a message with the isotope's
mass number and natural isotopic abundance. If the isotope is not a
naturally occuring isotope then a message will be displayed that the
isotope is not naturally occuring.
count_rate takes in both a measured and background spectra and a peak
energy and returns the count rate of the isotope.