Working Version of NAA_Analysis.py
Created function NAA_Analysis.py.
It asks the user to input the name of a sample measurement along with
relevant information.

It then finds all the peaks in the given spectrum and determines which
isotopes can correspond to each peak.

An excel spreadsheet is produced which contains all found energies along
with possible isotopes and their concentrations.