Improved peak net area calculation
Added function NAA_net_area, which calculates the net area at a given
energy. It first draws a line that's determined by the edges of the
Region of Interest (ROI), which generated by Full Width Half Max (FWHM).
It then acquires the area under the line by integrating. Then it
evaluates the net area by subtracting the total gross counts in the ROI
by the integral of the line. Uncertainties are generated and are
returned along with the net area.

Sub_regions are not required for this calculation since the line is
generated using the peak ROI.