Fixed calculation errors
Branching ratio was not being divided by 100 when the class was created.

Radioisotopes that do not have concentrations in the reference soil
sample were still being multiplied by the reference ratio. This caused a
systematic increase in Cs-137 concentrations and detection limits for
isotopes in the not_in_dirt list.