Added Cs134 compton regions as sub_regions input
Right region for Cs-134 made using Bi-214 609 keV peak. Makes a region
to the right of the 609 peak and uses that for the right compton region
for the Cs-134 along with the left side.

Also added the factor of sqrt(2) for when a compton region consists of
two regions rather than one.